Strings for Keys
Added stuff so that most keys are get saved as a string. Also added some
the start of some game stuff. There is a paddle at the top that can move
with the left and right arrow keys and every once and a while a ball
will appear and move towards the top of the screen.